refactoring util files to modules
