+++ src/models/user/utils/generate_hash.ts
@@ -1,8 +1,5 @@
 import bcrypt from 'bcryptjs';
-import { v4 as uuid } from 'uuid';
-import { SALT_ROUNDS } from '../../config/auth';
-
-export const generateKey = () => uuid().toString();
+import { SALT_ROUNDS } from '../../../config/auth';
 
 export const generateHash = (
   password: string,
@@ -16,17 +13,5 @@ export const generateHash = (
 
       callback(null, hash);
     });
-  });
-};
-
-export const compareHash = (
-  password: string,
-  hash: string,
-  callback: (err: Error, result?: boolean) => void,
-) => {
-  bcrypt.compare(password, hash, (err, result) => {
-    if (err) return callback(err);
-
-    callback(null, result);
   });
 };
