+++ src/models/user/utils/generate_key.ts
@@ -1,0 +1,3 @@
+import { v4 as uuid } from 'uuid';
+
+export const generateKey = () => uuid().toString();
