+++ src/validation/user/utils/validate_username.ts
@@ -1,5 +1,5 @@
-import { isMissing, compareType, ValidationError } from '../utils';
-import { User, UserInstance } from '../../models';
+import { isMissing, compareType, ValidationError } from '../../utils';
+import { User, UserInstance } from '../../../models';
 import { NativeError } from 'mongoose';
 
 type ValidateUsername = (
@@ -7,6 +7,7 @@ type ValidateUsername = (
   username: any,
   isRequired?: boolean,
 ) => Promise<void | ValidationError>;
+
 export const validateUsername: ValidateUsername = (username, isRequired = true) =>
   new Promise((resolve) => {
     if (isRequired && isMissing(username)) {
@@ -40,28 +41,4 @@ export const validateUsername: ValidateUsername = (username, isRequired = true) 
 
       resolve();
     });
-  });
-
-/* eslint-disable-next-line @typescript-eslint/no-explicit-any */
-type ValidatePassword = (password: any) => Promise<void | ValidationError>;
-export const validatePassword: ValidatePassword = (password) =>
-  new Promise((resolve) => {
-    if (isMissing(password)) {
-      return resolve('password is missing from input');
-    }
-
-    if (!compareType(password, 'string')) {
-      return resolve('password must be a string');
-    }
-
-    if (password.length < 8) {
-      return resolve('password must be at least 8 characters in length');
-    }
-
-    const regex = new RegExp('^(?=.*[a-z])(?=.*[A-Z])(?=.*\\d).+$');
-    if (!regex.test(password)) {
-      return resolve('password must have 1 uppercase, lowercase, and number character');
-    }
-
-    resolve();
   });
