+++ test/auth_endpoints/authenticate_token.spec.ts
@@ -1,7 +1,7 @@
 import assert from 'assert';
 import {
   getServerUrl,
-  connectDB,
+  connectDatabase,
   createTestUser,
   cleanupTestRecords,
   generateToken,
@@ -18,7 +18,7 @@ describe('[Auth] Authenticate Token', () => {
     let testUser: UserInstance;
 
     beforeAll(async () => {
-      connection = await connectDB();
+      connection = await connectDatabase();
       if (!connection) return 'could not connect to db';
 
       testUser = await createTestUser('Password1');
