+++ test/auth_endpoints/generate_token.spec.ts
@@ -1,7 +1,7 @@
 import assert from 'assert';
 import {
   getServerUrl,
-  connectDB,
+  connectDatabase,
   createTestUser,
   cleanupTestRecords,
   Connection,
@@ -17,7 +17,7 @@ describe('[Auth] Generate Token', () => {
     let testUser: UserInstance;
 
     beforeAll(async () => {
-      connection = await connectDB();
+      connection = await connectDatabase();
       if (!connection) return 'could not connect to db';
 
       testUser = await createTestUser('Password1');
